refactor simCLR.py
move train function to utils/train.py
